extract usePagination and useSort from useReferenceMany
